reworked in C++
Since 2006, my values and the values of the Python community
have diverged. So when I discovered I no longer had access
to a Python2 interpreter, porting it to Python3 was not an
option.

My current stance is, everything worthwhile can be written in
shell script or C++. So I picked C++.

Same functionality (except you can supply multiple quotes
files).